requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
opencv-python
numpy
python-multipart
segmentation-models-pytorch
scipy
